reportus/requirements.txt
pandas==2.0.3
python-dateutil==2.8.2
streamlit==1.25.0
